refactor: reorganize middleware structure and update routing to use new middleware definitions
